backend/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langgraph>=0.0.20
pymilvus>=2.3.6
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.3
python-docx>=1.1.0
docx2txt>=0.1.4
pypdf>=3.17.4
pywin32>=306
torch>=2.1.2
transformers>=4.36.2
sentence-transformers>=2.2.2
rank_bm25>=0.2.2
jieba>=0.42.1
python-multipart>=0.0.6
aiofiles>=23.2.1
tqdm>=4.66.0
